backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
transformers==4.36.0
torch>=2.0.0
accelerate==0.25.0
bitsandbytes==0.41.3
langchain==0.0.350
langchain-community==0.0.1
openai==1.3.7
textstat==0.7.3
nltk==3.8.1
spacy==3.7.2
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Optional: For TTS support
pyttsx3==2.90
gTTS==2.4.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
